app/bot/requirements.txt
nextcord==2.6.0
requests
psutil
python-dotenv
flask
PyYAML
qrcode
pytz
cryptography
PyJWT
psycopg2-binary
alembic
sqlalchemy
asyncpg
docker
py-cpuinfo
gputil
py3nvml
speedtest-cli
watchdog
redis==4.5.1
pytest
pytest-cov